shop_bot/requirements.txt
aiogram==3.7.0
aiosqlite==0.20.0
python-dotenv==1.0.0
Pillow==10.3.0
